Refactor: update getMe and change the routes
