+++ backend/src/middleware/auth.Middleware.ts
@@ -2,6 +2,7 @@ import jwt from "jsonwebtoken";
 import dotenv from "dotenv"
 
 import {Response, Request,NextFunction} from "express";
+
 dotenv.config()
 
 export const authentionToken = (req: Request, res: Response, next: NextFunction) => {
@@ -24,33 +25,3 @@ const token=authHeader && authHeader.split(" ")[1];
 
 }
 
-/*
-export const authenticateToken = (req: Request, res: Response, next: NextFunction) => {
-  const authHeader = req.headers["authorization"];
-  const token = authHeader && authHeader.split(" ")[1];
-
-  if (!token) {
- 
-    res.status(401).json({ mensagem: "Usuário não autenticado." });
-    return 
-  }
-
-  jwt.verify(token, process.env.JWT_SECRET as string, (err, user) => {
-    if (err) return res.status(403).json({ mensagem: "Token inválido" });
-
-    (req as any).user = user;
-    next();
-  });
-};
-
-export const authorizeRole = (role: string) => {
-  return (req: Request, res: Response, next: NextFunction) => {
-    const user = (req as any).user;
-
-    if (!user || user.role !== role) {
-      return res.status(403).json({ mensagem: "Acesso negado: permissão insuficiente." });
-    }
-
-    next();
-  };
-};*/
