+++ backend/src/routes/auth-routes.ts
@@ -1,0 +1,8 @@
+
+import Express from 'express'
+import { login } from '../controllers/auth.controller.ts'
+
+
+export const authRouter = Express.Router()
+
+authRouter.post('/login', login)
