+++ backend/src/server.ts
@@ -2,6 +2,7 @@ import express from 'express'
 import dotenv from 'dotenv'
 import mongoose from 'mongoose'
 import { router } from './routes/user.route.ts'
+import { authRouter } from './routes/auth-routes.ts'
 
 const app = express()
 dotenv.config()
@@ -10,6 +11,7 @@ app.use(express.json())
 const host = process.env.HOST || 'http://localhost'
 const port = process.env.PORT || 3000
 
+app.use('/auth', authRouter)
 app.use('/users', router)
